requirements.txt
aiosqlite==0.21.0
asyncpg==0.30.0
aiofiles==24.1.0
loguru==0.7.3

pydantic==2.10.6
pydantic-settings==2.8.1

aiogram==3.18.0

requests==2.32.3
python-dotenv==1.0.1
